// src/store/typed/modules/session/mutation-types.ts
export enum SessionMutationTypes {
    SET_USER = "SET_USER",
    CLEAR_SESSION = "CLEAR_SESSION",

    SET_SETTINGS = "SET_SETTINGS",
    SET_LANGUAGES = "SET_LANGUAGES",

    INITIALIZED = "INITIALIZED",

    COLLECTIONS = "COLLECTIONS",
    COLLECTION = "COLLECTION",

    SET_PLAYLISTS = "SET_PLAYLISTS",
    UPDATE_PLAYLIST = "UPDATE_PLAYLIST",
    SET_PLAYLIST = "SET_PLAYLIST",
    DELETE_PLAYLIST = "DELETE_PLAYLIST",
    
    EXTEND = "EXTEND",
    ERROR = "ERROR",
    REDIRECT = "REDIRECT",
    SET_LOG_ITEMS = "SET_LOG_ITEMS",
    CLEAR_LOGS = "CLEAR_LOGS"
}